refactor: billing pricer interface
This patch splits the monolithic ubp line into smaller pricers
allowing to compose the types going forward.

The patch also fixes an issue, that UBP FlatFee lines were considered
progressively billable (this is only an issue if the line is in arrears)
which is a rare use-case.
