+++ openmeter/billing/errors.go
@@ -32,13 +32,14 @@ var (
 	ErrFieldMustBePositive       = NewValidationError("field_must_be_positive", "field must be positive")
 	ErrFieldMustBePositiveOrZero = NewValidationError("field_must_be_positive_or_zero", "field must be positive or zero")
 
-	ErrInvoiceCannotAdvance      = NewValidationError("invoice_cannot_advance", "invoice cannot advance")
-	ErrInvoiceCannotBeEdited     = NewValidationError("invoice_cannot_be_edited", "invoice cannot be edited in the current state")
-	ErrInvoiceActionNotAvailable = NewValidationError("invoice_action_not_available", "invoice action not available")
-	ErrInvoiceLinesNotBillable   = NewValidationError("invoice_lines_not_billable", "invoice lines are not billable")
-	ErrInvoiceEmpty              = NewValidationError("invoice_empty", "invoice is empty")
-	ErrInvoiceNotFound           = NewValidationError("invoice_not_found", "invoice not found")
-	ErrInvoiceDeleteFailed       = NewValidationError("invoice_delete_failed", "invoice delete failed")
+	ErrInvoiceCannotAdvance                  = NewValidationError("invoice_cannot_advance", "invoice cannot advance")
+	ErrInvoiceCannotBeEdited                 = NewValidationError("invoice_cannot_be_edited", "invoice cannot be edited in the current state")
+	ErrInvoiceActionNotAvailable             = NewValidationError("invoice_action_not_available", "invoice action not available")
+	ErrInvoiceProgressiveBillingNotSupported = NewValidationError("invoice_progressive_billing_not_supported", "progressive billing is not supported")
+	ErrInvoiceLinesNotBillable               = NewValidationError("invoice_lines_not_billable", "invoice lines are not billable")
+	ErrInvoiceEmpty                          = NewValidationError("invoice_empty", "invoice is empty")
+	ErrInvoiceNotFound                       = NewValidationError("invoice_not_found", "invoice not found")
+	ErrInvoiceDeleteFailed                   = NewValidationError("invoice_delete_failed", "invoice delete failed")
 
 	ErrInvoiceLineFeatureHasNoMeters             = NewValidationError("invoice_line_feature_has_no_meters", "usage based invoice line: feature has no meters")
 	ErrInvoiceLineVolumeSplitNotSupported        = NewValidationError("invoice_line_graduated_split_not_supported", "graduated tiered pricing is not supported for split periods")
