+++ openmeter/billing/service/lineservice/usagebasedline.go
@@ -85,30 +85,24 @@ func (l usageBasedLine) CanBeInvoicedAsOf(ctx context.Context, in CanBeInvoicedA
 		return &l.line.Period, nil
 	}
 
-	if l.line.UsageBased.Price.Type() == productcatalog.TieredPriceType {
-		tiered, err := l.line.UsageBased.Price.AsTiered()
-		if err != nil {
-			return nil, err
-		}
-
-		if tiered.Mode == productcatalog.VolumeTieredPrice {
-			if l.line.ParentLine != nil {
-				if in.AsOf.Before(l.line.ParentLine.Period.End) {
-					return nil, nil
-				}
-
-				return &l.line.Period, nil
-			}
-
-			// Volume tiers are only billable if we have all the data acquired, as otherwise
-			// we might overcharge the customer (if we are at tier bundaries)
-			if in.AsOf.Before(l.line.Period.End) {
-				return nil, nil
-			}
-			return &l.line.Period, nil
-		}
+	// Progressive billing checks
+	pricer, err := l.getPricer()
+	if err != nil {
+		return nil, err
 	}
 
+	canBeInvoiced, err := pricer.CanBeInvoicedAsOf(l, in.AsOf)
+	if err != nil {
+		return nil, err
+	}
+
+	if !canBeInvoiced {
+		// If the pricer cannot be invoiced most probably due to the missing progressive billing support
+		// or invalid input, we should not bill the line
+		return nil, nil
+	}
+
+	// Let's check if the underlying meter can be billed in a progressive manner
 	meterAndFactory, err := l.service.resolveFeatureMeter(ctx, l.line.Namespace, l.line.UsageBased.FeatureKey)
 	if err != nil {
 		return nil, err
@@ -218,10 +212,7 @@ func (l *usageBasedLine) CalculateDetailedLines() error {
 		return fmt.Errorf("quantity and pre-line period quantity must be set for line[%s]", l.line.ID)
 	}
 
-	newDetailedLinesInput, err := l.calculateDetailedLines(&featureUsageResponse{
-		LinePeriodQty:    *l.line.UsageBased.Quantity,
-		PreLinePeriodQty: *l.line.UsageBased.PreLinePeriodQuantity,
-	})
+	newDetailedLinesInput, err := l.calculateDetailedLines()
 	if err != nil {
 		return err
 	}
@@ -241,80 +232,51 @@ func (l *usageBasedLine) CalculateDetailedLines() error {
 	return nil
 }
 
-func (l usageBasedLine) calculateDetailedLines(usage *featureUsageResponse) (newDetailedLinesInput, error) {
+func (l usageBasedLine) asPricerCalculateInput() (PricerCalculateInput, error) {
+	if l.line.UsageBased.Quantity == nil || l.line.UsageBased.PreLinePeriodQuantity == nil {
+		return PricerCalculateInput{}, fmt.Errorf("quantity and pre-line period quantity must be set for line[%s]", l.line.ID)
+	}
+
+	return PricerCalculateInput{
+		usageBasedLine:   l,
+		preLinePeriodQty: *l.line.UsageBased.PreLinePeriodQuantity,
+		linePeriodQty:    *l.line.UsageBased.Quantity,
+	}, nil
+}
+
+func (l usageBasedLine) getPricer() (Pricer, error) {
+	var pricer Pricer
+
 	switch l.line.UsageBased.Price.Type() {
 	case productcatalog.FlatPriceType:
-		flatPrice, err := l.line.UsageBased.Price.AsFlat()
-		if err != nil {
-			return nil, fmt.Errorf("converting price to flat price: %w", err)
-		}
-		return l.calculateFlatPriceDetailedLines(usage, flatPrice)
-
+		pricer = flatPricer{}
 	case productcatalog.UnitPriceType:
-		unitPrice, err := l.line.UsageBased.Price.AsUnit()
-		if err != nil {
-			return nil, fmt.Errorf("converting price to unit price: %w", err)
-		}
-
-		return l.calculateUnitPriceDetailedLines(usage, unitPrice)
-	case productcatalog.DynamicPriceType:
-		dynamicPrice, err := l.line.UsageBased.Price.AsDynamic()
-		if err != nil {
-			return nil, fmt.Errorf("converting price to dynamic price: %w", err)
-		}
-		return l.calculateDynamicPriceDetailedLines(usage, dynamicPrice)
-	case productcatalog.PackagePriceType:
-		packagePrice, err := l.line.UsageBased.Price.AsPackage()
-		if err != nil {
-			return nil, fmt.Errorf("converting price to package price: %w", err)
-		}
-		return l.calculatePackagePriceDetailedLines(usage, packagePrice)
+		pricer = unitPricer{}
 	case productcatalog.TieredPriceType:
-		tieredPrice, err := l.line.UsageBased.Price.AsTiered()
-		if err != nil {
-			return nil, fmt.Errorf("converting price to tiered price: %w", err)
-		}
-
-		switch tieredPrice.Mode {
-		case productcatalog.VolumeTieredPrice:
-			return l.calculateVolumeTieredPriceDetailedLines(usage, tieredPrice)
-
-		case productcatalog.GraduatedTieredPrice:
-			return l.calculateGraduatedTieredPriceDetailedLines(usage, tieredPrice)
-		default:
-			return nil, fmt.Errorf("unsupported tiered price mode: %s", tieredPrice.Mode)
-		}
+		pricer = tieredPricer{}
+	case productcatalog.PackagePriceType:
+		pricer = packagePricer{}
+	case productcatalog.DynamicPriceType:
+		pricer = dynamicPricer{}
 	default:
 		return nil, fmt.Errorf("unsupported price type: %s", l.line.UsageBased.Price.Type())
 	}
+
+	return pricer, nil
 }
 
-func (l usageBasedLine) calculateFlatPriceDetailedLines(_ *featureUsageResponse, flatPrice productcatalog.FlatPrice) (newDetailedLinesInput, error) {
-	// Flat price is the same as the non-metered version, we just allow attaching entitlements to it
-	switch {
-	case flatPrice.PaymentTerm == productcatalog.InAdvancePaymentTerm && l.IsFirstInPeriod():
-		return newDetailedLinesInput{
-			{
-				Name:                   l.line.Name,
-				Quantity:               alpacadecimal.NewFromFloat(1),
-				PerUnitAmount:          flatPrice.Amount,
-				ChildUniqueReferenceID: FlatPriceChildUniqueReferenceID,
-				PaymentTerm:            productcatalog.InAdvancePaymentTerm,
-			},
-		}, nil
-	case flatPrice.PaymentTerm != productcatalog.InAdvancePaymentTerm && l.IsLastInPeriod():
-		return newDetailedLinesInput{
-			{
-				Name:                   l.line.Name,
-				Quantity:               alpacadecimal.NewFromFloat(1),
-				PerUnitAmount:          flatPrice.Amount,
-				ChildUniqueReferenceID: FlatPriceChildUniqueReferenceID,
-				PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-			},
-		}, nil
+func (l usageBasedLine) calculateDetailedLines() (newDetailedLinesInput, error) {
+	pricer, err := l.getPricer()
+	if err != nil {
+		return nil, err
 	}
 
-	return nil, nil
+	pricerInput, err := l.asPricerCalculateInput()
+	if err != nil {
+		return nil, err
+	}
+
+	return pricer.Calculate(pricerInput)
 }
 
 type applyCommitmentsInput struct {
@@ -388,517 +350,6 @@ func (l usageBasedLine) applyCommitments(in applyCommitmentsInput) (newDetailedL
 	}
 
 	return in.DetailedLines, nil
-}
-
-func (l usageBasedLine) calculateUnitPriceDetailedLines(usage *featureUsageResponse, unitPrice productcatalog.UnitPrice) (newDetailedLinesInput, error) {
-	var out newDetailedLinesInput
-
-	if usage.LinePeriodQty.IsPositive() {
-		out = newDetailedLinesInput{
-			{
-				Name:                   fmt.Sprintf("%s: usage in period", l.line.Name),
-				Quantity:               usage.LinePeriodQty,
-				PerUnitAmount:          unitPrice.Amount,
-				ChildUniqueReferenceID: UnitPriceUsageChildUniqueReferenceID,
-				PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-			},
-		}
-	}
-
-	amountBilledInPreviousPeriods := l.currency.RoundToPrecision(usage.PreLinePeriodQty.Mul(unitPrice.Amount))
-
-	detailedLines, err := l.applyCommitments(applyCommitmentsInput{
-		Commitments:                   unitPrice.Commitments,
-		DetailedLines:                 out,
-		AmountBilledInPreviousPeriods: amountBilledInPreviousPeriods,
-		MinimumSpendReferenceID:       UnitPriceMinSpendChildUniqueReferenceID,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return detailedLines, nil
-}
-
-func (l usageBasedLine) calculateDynamicPriceDetailedLines(usage *featureUsageResponse, dynamicPrice productcatalog.DynamicPrice) (newDetailedLinesInput, error) {
-	var out newDetailedLinesInput
-
-	if usage.LinePeriodQty.IsPositive() {
-		amountInPeriod := l.currency.RoundToPrecision(
-			usage.LinePeriodQty.Mul(dynamicPrice.Multiplier),
-		)
-
-		out = newDetailedLinesInput{
-			{
-				Name:                   fmt.Sprintf("%s: usage in period", l.line.Name),
-				Quantity:               alpacadecimal.NewFromInt(1),
-				PerUnitAmount:          amountInPeriod,
-				ChildUniqueReferenceID: UsageChildUniqueReferenceID,
-				PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-			},
-		}
-	}
-
-	amountBilledInPreviousPeriods := l.currency.RoundToPrecision(usage.PreLinePeriodQty.Mul(dynamicPrice.Multiplier))
-
-	detailedLines, err := l.applyCommitments(applyCommitmentsInput{
-		Commitments:                   dynamicPrice.Commitments,
-		DetailedLines:                 out,
-		AmountBilledInPreviousPeriods: amountBilledInPreviousPeriods,
-		MinimumSpendReferenceID:       MinSpendChildUniqueReferenceID,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return detailedLines, nil
-}
-
-func (l usageBasedLine) getNumberOfPackages(qty alpacadecimal.Decimal, packageSize alpacadecimal.Decimal) alpacadecimal.Decimal {
-	requiredPackages := qty.Div(packageSize).Floor()
-
-	if qty.Mod(packageSize).IsZero() {
-		return requiredPackages
-	}
-
-	return requiredPackages.Add(DecimalOne)
-}
-
-func (l usageBasedLine) calculatePackagePriceDetailedLines(usage *featureUsageResponse, packagePrice productcatalog.PackagePrice) (newDetailedLinesInput, error) {
-	var out newDetailedLinesInput
-
-	totalUsage := usage.LinePeriodQty.Add(usage.PreLinePeriodQty)
-
-	preLinePeriodPackages := l.getNumberOfPackages(usage.PreLinePeriodQty, packagePrice.QuantityPerPackage)
-	if l.IsFirstInPeriod() {
-		preLinePeriodPackages = alpacadecimal.Zero
-	}
-
-	postLinePeriodPackages := l.getNumberOfPackages(totalUsage, packagePrice.QuantityPerPackage)
-
-	toBeBilledPackages := postLinePeriodPackages.Sub(preLinePeriodPackages)
-
-	if !toBeBilledPackages.IsZero() {
-		out = newDetailedLinesInput{
-			{
-				Name:                   fmt.Sprintf("%s: usage in period", l.line.Name),
-				Quantity:               toBeBilledPackages,
-				PerUnitAmount:          packagePrice.Amount,
-				ChildUniqueReferenceID: UsageChildUniqueReferenceID,
-				PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-			},
-		}
-	}
-
-	detailedLines, err := l.applyCommitments(applyCommitmentsInput{
-		Commitments:                   packagePrice.Commitments,
-		DetailedLines:                 out,
-		AmountBilledInPreviousPeriods: l.currency.RoundToPrecision(preLinePeriodPackages.Mul(packagePrice.Amount)),
-		MinimumSpendReferenceID:       MinSpendChildUniqueReferenceID,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return detailedLines, nil
-}
-
-func (l usageBasedLine) calculateVolumeTieredPriceDetailedLines(usage *featureUsageResponse, price productcatalog.TieredPrice) (newDetailedLinesInput, error) {
-	if !usage.PreLinePeriodQty.IsZero() {
-		return nil, billing.ErrInvoiceLineVolumeSplitNotSupported
-	}
-
-	if !l.IsLastInPeriod() {
-		return nil, nil
-	}
-
-	out := make(newDetailedLinesInput, 0, 4)
-
-	findTierRes, err := findTierForQuantity(price, usage.LinePeriodQty)
-	if err != nil {
-		return nil, err
-	}
-
-	tier := findTierRes.Tier
-	tierIndex := findTierRes.Index
-
-	if tier.FlatPrice != nil {
-		line := newDetailedLineInput{
-			Name:                   fmt.Sprintf("%s: flat price for tier %d", l.line.Name, tierIndex+1),
-			Quantity:               alpacadecimal.NewFromFloat(1),
-			PerUnitAmount:          tier.FlatPrice.Amount,
-			ChildUniqueReferenceID: VolumeFlatPriceChildUniqueReferenceID,
-			PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-		}
-
-		if price.MaximumAmount != nil {
-			line = line.AddDiscountForOverage(addDiscountInput{
-				BilledAmountBeforeLine: out.Sum(l.currency),
-				MaxSpend:               *price.MaximumAmount,
-				Currency:               l.currency,
-			})
-		}
-		out = append(out, line)
-	}
-
-	if tier.UnitPrice != nil && !usage.LinePeriodQty.IsZero() {
-		line := newDetailedLineInput{
-			Name:                   fmt.Sprintf("%s: unit price for tier %d", l.line.Name, tierIndex+1),
-			Quantity:               usage.LinePeriodQty,
-			PerUnitAmount:          tier.UnitPrice.Amount,
-			ChildUniqueReferenceID: VolumeUnitPriceChildUniqueReferenceID,
-			PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-		}
-
-		if price.MaximumAmount != nil {
-			line = line.AddDiscountForOverage(addDiscountInput{
-				BilledAmountBeforeLine: out.Sum(l.currency),
-				MaxSpend:               *price.MaximumAmount,
-				Currency:               l.currency,
-			})
-		}
-
-		out = append(out, line)
-	}
-
-	total := out.Sum(l.currency)
-
-	if price.MinimumAmount != nil {
-		normalizedMinimumAmount := l.currency.RoundToPrecision(*price.MinimumAmount)
-
-		if total.LessThan(normalizedMinimumAmount) {
-			out = append(out, newDetailedLineInput{
-				Name:                   fmt.Sprintf("%s: minimum spend", l.line.Name),
-				Quantity:               alpacadecimal.NewFromFloat(1),
-				PerUnitAmount:          normalizedMinimumAmount.Sub(total),
-				ChildUniqueReferenceID: VolumeMinSpendChildUniqueReferenceID,
-				PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-				Category:               billing.FlatFeeCategoryCommitment,
-			})
-		}
-	}
-
-	return out, nil
-}
-
-type findTierForQuantityResult struct {
-	Tier  *productcatalog.PriceTier
-	Index int
-}
-
-func findTierForQuantity(price productcatalog.TieredPrice, quantity alpacadecimal.Decimal) (findTierForQuantityResult, error) {
-	for i, tier := range price.WithSortedTiers().Tiers {
-		if tier.UpToAmount == nil || quantity.LessThanOrEqual(*tier.UpToAmount) {
-			return findTierForQuantityResult{
-				Tier:  &price.Tiers[i],
-				Index: i,
-			}, nil
-		}
-	}
-
-	// Technically this should not happen, as the last tier should have an upper limit of infinity
-	return findTierForQuantityResult{}, fmt.Errorf("could not find tier for quantity %s: %w", quantity, billing.ErrInvoiceLineMissingOpenEndedTier)
-}
-
-func (l usageBasedLine) calculateGraduatedTieredPriceDetailedLines(usage *featureUsageResponse, price productcatalog.TieredPrice) (newDetailedLinesInput, error) {
-	out := make(newDetailedLinesInput, 0, len(price.Tiers))
-
-	err := tieredPriceCalculator(tieredPriceCalculatorInput{
-		TieredPrice: price,
-		FromQty:     usage.PreLinePeriodQty,
-		ToQty:       usage.LinePeriodQty.Add(usage.PreLinePeriodQty),
-		Currency:    l.currency,
-		TierCallbackFn: func(in tierCallbackInput) error {
-			billedAmount := in.PreviousTotalAmount
-
-			tierIndex := in.TierIndex + 1
-
-			if in.Tier.UnitPrice != nil && in.Quantity.IsPositive() {
-				newLine := newDetailedLineInput{
-					Name:                   fmt.Sprintf("%s: usage price for tier %d", l.line.Name, tierIndex),
-					Quantity:               in.Quantity,
-					PerUnitAmount:          in.Tier.UnitPrice.Amount,
-					ChildUniqueReferenceID: fmt.Sprintf(GraduatedTieredPriceUsageChildUniqueReferenceID, tierIndex),
-					PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-				}
-
-				if price.MaximumAmount != nil {
-					newLine = newLine.AddDiscountForOverage(addDiscountInput{
-						BilledAmountBeforeLine: billedAmount,
-						MaxSpend:               *price.MaximumAmount,
-						Currency:               l.currency,
-					})
-				}
-
-				billedAmount = billedAmount.Add(in.Quantity.Mul(in.Tier.UnitPrice.Amount))
-
-				out = append(out, newLine)
-			}
-
-			// If have already billed this flat price for the previous split line, so we can skip it
-			shouldFirstFlatLineBeBilled := in.TierIndex > 0 || l.IsFirstInPeriod()
-
-			// Flat price is always billed for the whole tier when we are crossing the tier boundary
-			if in.Tier.FlatPrice != nil && in.AtTierBoundary && shouldFirstFlatLineBeBilled {
-				newLine := newDetailedLineInput{
-					Name:                   fmt.Sprintf("%s: flat price for tier %d", l.line.Name, tierIndex),
-					Quantity:               alpacadecimal.NewFromFloat(1),
-					PerUnitAmount:          in.Tier.FlatPrice.Amount,
-					ChildUniqueReferenceID: fmt.Sprintf(GraduatedTieredFlatPriceChildUniqueReferenceID, tierIndex),
-					PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-				}
-
-				if price.MaximumAmount != nil {
-					newLine = newLine.AddDiscountForOverage(addDiscountInput{
-						BilledAmountBeforeLine: billedAmount,
-						MaxSpend:               *price.MaximumAmount,
-						Currency:               l.currency,
-					})
-				}
-
-				out = append(out, newLine)
-			}
-			return nil
-		},
-		FinalizerFn: func(periodTotal alpacadecimal.Decimal) error {
-			if l.IsLastInPeriod() && price.MinimumAmount != nil {
-				normalizedMinimumAmount := l.currency.RoundToPrecision(*price.MinimumAmount)
-
-				if periodTotal.LessThan(normalizedMinimumAmount) {
-					out = append(out, newDetailedLineInput{
-						Name:                   fmt.Sprintf("%s: minimum spend", l.line.Name),
-						Quantity:               alpacadecimal.NewFromFloat(1),
-						PerUnitAmount:          normalizedMinimumAmount.Sub(periodTotal),
-						ChildUniqueReferenceID: GraduatedMinSpendChildUniqueReferenceID,
-						PaymentTerm:            productcatalog.InArrearsPaymentTerm,
-						Category:               billing.FlatFeeCategoryCommitment,
-					})
-				}
-			}
-
-			return nil
-		},
-	})
-	if err != nil {
-		return nil, fmt.Errorf("calculating tiered price: %w", err)
-	}
-
-	return out, nil
-}
-
-type tierRange struct {
-	Tier      productcatalog.PriceTier
-	TierIndex int
-
-	FromQty        alpacadecimal.Decimal // exclusive
-	ToQty          alpacadecimal.Decimal // inclusive
-	AtTierBoundary bool
-}
-
-type tierCallbackInput struct {
-	Tier                productcatalog.PriceTier
-	TierIndex           int
-	Quantity            alpacadecimal.Decimal
-	AtTierBoundary      bool
-	PreviousTotalAmount alpacadecimal.Decimal
-}
-
-type tieredPriceCalculatorInput struct {
-	TieredPrice productcatalog.TieredPrice
-	// FromQty is the quantity that was already billed for the previous tiers (exclusive)
-	FromQty alpacadecimal.Decimal
-	// ToQty is the quantity that we are going to bill for this tiered price (inclusive)
-	ToQty alpacadecimal.Decimal
-
-	Currency currencyx.Calculator
-
-	TierCallbackFn     func(tierCallbackInput) error
-	FinalizerFn        func(total alpacadecimal.Decimal) error
-	IntrospectRangesFn func(ranges []tierRange)
-}
-
-func (i tieredPriceCalculatorInput) Validate() error {
-	if err := i.TieredPrice.Validate(); err != nil {
-		return err
-	}
-
-	if i.TieredPrice.Mode != productcatalog.GraduatedTieredPrice {
-		return fmt.Errorf("only graduated tiered prices are supported")
-	}
-
-	if i.FromQty.IsNegative() {
-		return fmt.Errorf("from quantity must be zero or positive")
-	}
-
-	if i.ToQty.IsNegative() {
-		return fmt.Errorf("to quantity must be zero or positive")
-	}
-
-	if i.ToQty.LessThan(i.FromQty) {
-		return fmt.Errorf("to quantity must be greater or equal to from quantity")
-	}
-
-	if i.Currency.Currency == "" {
-		return fmt.Errorf("currency calculator is required")
-	}
-
-	return nil
-}
-
-func splitTierRangeAtBoundary(from, to alpacadecimal.Decimal, qtyRange tierRange) []tierRange {
-	res := make([]tierRange, 0, 3)
-
-	// Pending line is always the last line, as we might need to split it
-	pendingLine := qtyRange
-
-	// If from == in.FromQty we don't need to split the range, as the range is already at some boundary
-	if pendingLine.FromQty.LessThan(from) && pendingLine.ToQty.GreaterThan(from) {
-		// We need to split the range at the from boundary
-		res = append(res, tierRange{
-			Tier:      pendingLine.Tier,
-			TierIndex: pendingLine.TierIndex,
-
-			FromQty: pendingLine.FromQty,
-			ToQty:   from,
-
-			AtTierBoundary: pendingLine.AtTierBoundary,
-		})
-
-		pendingLine = tierRange{
-			Tier:      pendingLine.Tier,
-			TierIndex: pendingLine.TierIndex,
-
-			FromQty: from,
-			ToQty:   pendingLine.ToQty,
-		}
-	}
-
-	// If to == in.ToQty we don't need to split the range, as the range is already at some boundary
-	if pendingLine.FromQty.LessThan(to) && pendingLine.ToQty.GreaterThan(to) {
-		res = append(res, tierRange{
-			Tier:      pendingLine.Tier,
-			TierIndex: pendingLine.TierIndex,
-
-			FromQty: pendingLine.FromQty,
-			ToQty:   to,
-
-			AtTierBoundary: pendingLine.AtTierBoundary,
-		})
-		pendingLine = tierRange{
-			Tier:      pendingLine.Tier,
-			TierIndex: pendingLine.TierIndex,
-
-			FromQty: to,
-			ToQty:   pendingLine.ToQty,
-		}
-	}
-
-	return append(res, pendingLine)
-}
-
-// getTotalAmountForGraduatedTieredPrice calculates the total amount for a graduated tiered price for a given quantity
-// without considering any discounts
-func tieredPriceCalculator(in tieredPriceCalculatorInput) error {
-	// Note: this is not the most efficient algorithm, but it is at least pseudo-readable
-	if err := in.Validate(); err != nil {
-		return err
-	}
-
-	// Let's break up the tiers and the input data into a sequence of periods, for easier processing
-	// Invariant of the qtyRanges:
-	// - Non overlapping ranges
-	// - The ranges are sorted by the from quantity
-	// - There is always one range for which range.From == in.FromQty
-	// - There is always one range for which range.ToQty == in.ToQty
-	qtyRanges := make([]tierRange, 0, len(in.TieredPrice.Tiers)+2)
-
-	previousTierQty := alpacadecimal.Zero
-	for idx, tier := range in.TieredPrice.WithSortedTiers().Tiers {
-		if previousTierQty.GreaterThanOrEqual(in.ToQty) {
-			// We already have enough data to bill for this tiered price
-			break
-		}
-
-		// Given that the previous tier's max qty was less than then in.ToQty, toQty will fall into the
-		// open ended tier, so we can safely use it as the upper bound
-		tierUpperBound := in.ToQty
-		if tier.UpToAmount != nil {
-			tierUpperBound = *tier.UpToAmount
-		}
-
-		input := tierRange{
-			Tier:           tier,
-			TierIndex:      idx,
-			AtTierBoundary: true,
-			FromQty:        previousTierQty,
-			ToQty:          tierUpperBound,
-		}
-
-		qtyRanges = append(qtyRanges, splitTierRangeAtBoundary(in.FromQty, in.ToQty, input)...)
-
-		previousTierQty = tierUpperBound
-	}
-
-	if in.ToQty.Equal(alpacadecimal.Zero) {
-		// We need to add the first range, in case there's a flat price component
-		qtyRanges = append(qtyRanges, tierRange{
-			Tier:           in.TieredPrice.Tiers[0],
-			TierIndex:      0,
-			AtTierBoundary: true,
-			FromQty:        alpacadecimal.Zero,
-			ToQty:          alpacadecimal.Zero,
-		})
-	}
-
-	if in.IntrospectRangesFn != nil {
-		in.IntrospectRangesFn(qtyRanges)
-	}
-
-	// Now that we have the ranges, let's iterate over the ranges and calculate the cummulative total amount
-	// and call the callback for each in-scope range
-	total := alpacadecimal.Zero
-	shouldEmitCallbacks := false
-	for _, qtyRange := range qtyRanges {
-		if qtyRange.FromQty.Equal(in.FromQty) {
-			shouldEmitCallbacks = true
-		}
-
-		if shouldEmitCallbacks && in.TierCallbackFn != nil {
-			err := in.TierCallbackFn(tierCallbackInput{
-				Tier:                qtyRange.Tier,
-				TierIndex:           qtyRange.TierIndex,
-				Quantity:            qtyRange.ToQty.Sub(qtyRange.FromQty),
-				PreviousTotalAmount: total,
-				AtTierBoundary:      qtyRange.AtTierBoundary,
-			})
-			if err != nil {
-				return err
-			}
-		}
-
-		// Let's update totals
-		if qtyRange.Tier.FlatPrice != nil && qtyRange.AtTierBoundary {
-			total = total.Add(in.Currency.RoundToPrecision(qtyRange.Tier.FlatPrice.Amount))
-		}
-
-		if qtyRange.Tier.UnitPrice != nil {
-			total = total.Add(in.Currency.RoundToPrecision(qtyRange.ToQty.Sub(qtyRange.FromQty).Mul(qtyRange.Tier.UnitPrice.Amount)))
-		}
-
-		// We should only calculate totals up to in.ToQty (given tiers are open-ended we cannot have a full upper bound
-		// either ways)
-		if qtyRange.ToQty.GreaterThanOrEqual(in.ToQty) {
-			break
-		}
-	}
-
-	if in.FinalizerFn != nil {
-		if err := in.FinalizerFn(total); err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 type newDetailedLinesInput []newDetailedLineInput
